[feature] シーン遷移後、迷路作成実装
