+++ Assets/Scenes/InvasionPhase.unity
@@ -576,7 +576,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 91991389372cd9944a8eaf505baf9651, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  tilePrefab: {fileID: 2227749310468335157, guid: 26a8b4975ac1871409a5e906e043cdeb, type: 3}
+  createPhaseTilePrefab: {fileID: -5516297106693624721, guid: abd90deeafc55a44ea812a9e246a2a13, type: 3}
 --- !u!1 &1522057559
 GameObject:
   m_ObjectHideFlags: 0
