+++ Assets/Scripts/ATile.cs
@@ -2,7 +2,7 @@ using Enums;
 using Traps;
 using UnityEngine;
 
-public class ATile : MonoBehaviour
+public abstract class ATile : MonoBehaviour
 {
     /** デフォルトのモデル */
     [SerializeField] protected Mesh defaultModel;
@@ -189,5 +189,26 @@ public class ATile : MonoBehaviour
     public void ResetTile()
     {
         Destroy(gameObject);
+    }
+
+    /**
+     * タイルを空に設定する
+     */
+    public void SetNone()
+    {
+        // タイルの種類を空に設定
+        TileType = TileTypes.Nothing;
+
+        // モデルを変更
+        var meshFilter = GetComponent<MeshFilter>();
+        meshFilter.mesh = defaultModel;
+
+        // アウトラインを消す
+        var outline = GetComponent<Outline>();
+        if (outline) outline.enabled = false;
+
+
+        // 回転を元に戻す
+        transform.rotation = Quaternion.Euler(-90, 0, 0);
     }
 }
