+++ Assets/Scripts/ATile.cs.meta
@@ -1,3 +1,18 @@
 fileFormatVersion: 2
 guid: e2b47c3500764bf393e73dccfc5ab71e
-timeCreated: 1720626515
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - defaultModel: {fileID: -3084497227037556731, guid: ab0a6f60044f8d14192971310666449f, type: 3}
+  - noneModel: {fileID: -5776783431495031102, guid: 9fd84ac705ca4cb4a96a26a9413fc5f0, type: 3}
+  - deadEndModel: {fileID: -8215893812795707571, guid: 309f198e7bafee2459d634e1f5b7bf7e, type: 3}
+  - straightModel: {fileID: -2405365958620393163, guid: 075776c8ba67a7b42b66b053e74f73a3, type: 3}
+  - cornerModel: {fileID: 4932170253921981256, guid: 0be1faee2b7ffd8479cdff2de7072011, type: 3}
+  - tJunctionModel: {fileID: 6750382564028958556, guid: 68136e2c800b0664189b53e07fec4ee2, type: 3}
+  - crossroadsModel: {fileID: -7532876345782023577, guid: f97e0ca8eb6a9044b8c25f71ba8cb863, type: 3}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
