+++ Assets/Scripts/InvasionPhase/InvasionMazeController.cs
@@ -1,6 +1,5 @@
-using System.Collections.Generic;
-using CreatePhase;
 using DataClass;
+using Enums;
 using UnityEngine;
 using UnityEngine.Serialization;
 
@@ -10,12 +9,12 @@ namespace InvasionPhase
     {
         /** タイルのプレハブ */
         [FormerlySerializedAs("tilePrefab")] [SerializeField]
-        private CreatePhaseTile createPhaseTilePrefab;
+        private InvasionPhaseTile createPhaseTilePrefab;
 
         private Vector3 _mazeOrigin;
 
         /** 迷路のデータ */
-        private CreatePhaseTile[][] Maze { get; set; }
+        private InvasionPhaseTile[][] Maze { get; set; }
 
         public void Create(TileData[][] tiles, TrapData[] trapData)
         {
@@ -26,18 +25,22 @@ namespace InvasionPhase
             _mazeOrigin = new Vector3(-(mazeColumns - 1) / 2.0f, 0, -(mazeRows - 1) / 2.0f);
             // すべてのタイルを生成し、初期化する
             // 行の初期化
-            Maze = new CreatePhaseTile[mazeRows][];
+            Maze = new InvasionPhaseTile[mazeRows][];
             for (var row = 0; row < mazeRows; row++)
             {
                 // 列の初期化
-                Maze[row] = new CreatePhaseTile[mazeColumns];
+                Maze[row] = new InvasionPhaseTile[mazeColumns];
                 for (var column = 0; column < mazeColumns; column++)
                 {
+                    var tileData = tiles[row][column];
+
                     // タイルの位置と回転を設定
                     var tilePosition = new Vector3(column, 0, row) * Environment.TileSize + _mazeOrigin;
                     var tileRotation = Quaternion.Euler(-90, 0, 0);
                     // タイルを生成し、初期化する
                     var newTile = Instantiate(createPhaseTilePrefab, tilePosition, tileRotation);
+
+                    newTile.Initialize(row, column, tileData.TileType, tileData.RoadAdjust);
 
                     // タイルを迷路に追加する
                     Maze[row][column] = newTile;
